reorg: move config creation to main package
