+++ collector/engine_innodb.go
@@ -103,5 +103,5 @@ func (*ScrapeEngineInnodbStatus) Scrape(ctx context.Context, db *sql.DB, ch chan
 var scrapeEngineInnodbStatus Scraper = &ScrapeEngineInnodbStatus{}
 
 func init() {
-	mustRegisterWithDefaults(scrapeEngineInnodbStatus, false)
+	mustRegisterScraperWithDefaults(scrapeEngineInnodbStatus, false)
 }
