+++ collector/global_status.go
@@ -229,5 +229,5 @@ func (*ScrapeGlobalStatus) Scrape(ctx context.Context, db *sql.DB, ch chan<- pro
 var scrapeGlobalStatus Scraper = &ScrapeGlobalStatus{}
 
 func init() {
-	mustRegisterWithDefaults(scrapeGlobalStatus, true)
+	mustRegisterScraperWithDefaults(scrapeGlobalStatus, true)
 }
