+++ collector/info_schema_auto_increment.go
@@ -104,5 +104,5 @@ func (*ScrapeAutoIncrementColumns) Scrape(ctx context.Context, db *sql.DB, ch ch
 var scrapeAutoIncrementColumns Scraper = &ScrapeAutoIncrementColumns{}
 
 func init() {
-	mustRegisterWithDefaults(scrapeAutoIncrementColumns, false)
+	mustRegisterScraperWithDefaults(scrapeAutoIncrementColumns, false)
 }
