+++ collector/info_schema_innodb_cmpmem.go
@@ -103,5 +103,5 @@ func (*ScrapeInnodbCmpMem) Scrape(ctx context.Context, db *sql.DB, ch chan<- pro
 var scrapeInnodbCmpMem Scraper = &ScrapeInnodbCmpMem{}
 
 func init() {
-	mustRegisterWithDefaults(scrapeInnodbCmpMem, true)
+	mustRegisterScraperWithDefaults(scrapeInnodbCmpMem, true)
 }
