+++ collector/info_schema_innodb_metrics.go
@@ -203,5 +203,5 @@ func (*ScrapeInnodbMetrics) Scrape(ctx context.Context, db *sql.DB, ch chan<- pr
 var scrapeInnodbMetrics Scraper = &ScrapeInnodbMetrics{}
 
 func init() {
-	mustRegisterWithDefaults(scrapeInnodbMetrics, false)
+	mustRegisterScraperWithDefaults(scrapeInnodbMetrics, false)
 }
