+++ collector/info_schema_query_response_time.go
@@ -146,5 +146,5 @@ func (*ScrapeQueryResponseTime) Scrape(ctx context.Context, db *sql.DB, ch chan<
 var scrapeQueryResponseTime Scraper = &ScrapeQueryResponseTime{}
 
 func init() {
-	mustRegisterWithDefaults(scrapeQueryResponseTime, false)
+	mustRegisterScraperWithDefaults(scrapeQueryResponseTime, false)
 }
