+++ collector/info_schema_schemastats.go
@@ -129,5 +129,5 @@ func (*ScrapeSchemaStat) Scrape(ctx context.Context, db *sql.DB, ch chan<- prome
 var scrapeSchemaStat Scraper = &ScrapeSchemaStat{}
 
 func init() {
-	mustRegisterWithDefaults(scrapeSchemaStat, false)
+	mustRegisterScraperWithDefaults(scrapeSchemaStat, false)
 }
