+++ collector/info_schema_tablestats.go
@@ -129,5 +129,5 @@ func (*ScrapeTableStat) Scrape(ctx context.Context, db *sql.DB, ch chan<- promet
 var scrapeTableStat Scraper = &ScrapeTableStat{}
 
 func init() {
-	mustRegisterWithDefaults(scrapeTableStat, false)
+	mustRegisterScraperWithDefaults(scrapeTableStat, false)
 }
