+++ collector/info_schema_userstats.go
@@ -219,5 +219,5 @@ func (*ScrapeUserStat) Scrape(ctx context.Context, db *sql.DB, ch chan<- prometh
 var scrapeUserStat Scraper = &ScrapeUserStat{}
 
 func init() {
-	mustRegisterWithDefaults(scrapeUserStat, false)
+	mustRegisterScraperWithDefaults(scrapeUserStat, false)
 }
