+++ collector/mysql_user.go
@@ -295,5 +295,5 @@ func (s *ScrapeUser) Scrape(ctx context.Context, db *sql.DB, ch chan<- prometheu
 var scrapeUser Scraper = &ScrapeUser{}
 
 func init() {
-	mustRegisterWithDefaults(scrapeUser, false)
+	mustRegisterScraperWithDefaults(scrapeUser, false)
 }
