+++ collector/perf_schema_events_waits.go
@@ -96,5 +96,5 @@ func (*ScrapePerfEventsWaits) Scrape(ctx context.Context, db *sql.DB, ch chan<- 
 var scrapePerfEventsWaits Scraper = &ScrapePerfEventsWaits{}
 
 func init() {
-	mustRegisterWithDefaults(scrapePerfEventsWaits, false)
+	mustRegisterScraperWithDefaults(scrapePerfEventsWaits, false)
 }
