+++ collector/perf_schema_file_events.go
@@ -133,5 +133,5 @@ func (*ScrapePerfFileEvents) Scrape(ctx context.Context, db *sql.DB, ch chan<- p
 var scrapePerfFileEvents Scraper = &ScrapePerfFileEvents{}
 
 func init() {
-	mustRegisterWithDefaults(scrapePerfFileEvents, false)
+	mustRegisterScraperWithDefaults(scrapePerfFileEvents, false)
 }
