+++ collector/perf_schema_file_instances.go
@@ -171,5 +171,5 @@ func (s *ScrapePerfFileInstances) Scrape(ctx context.Context, db *sql.DB, ch cha
 var scrapePerfFileInstances Scraper = &ScrapePerfFileInstances{}
 
 func init() {
-	mustRegisterWithDefaults(scrapePerfFileInstances, false)
+	mustRegisterScraperWithDefaults(scrapePerfFileInstances, false)
 }
