+++ collector/perf_schema_index_io_waits.go
@@ -131,5 +131,5 @@ func (*ScrapePerfIndexIOWaits) Scrape(ctx context.Context, db *sql.DB, ch chan<-
 var scrapePerfIndexIOWaits Scraper = &ScrapePerfIndexIOWaits{}
 
 func init() {
-	mustRegisterWithDefaults(scrapePerfIndexIOWaits, false)
+	mustRegisterScraperWithDefaults(scrapePerfIndexIOWaits, false)
 }
