+++ collector/perf_schema_replication_group_members.go
@@ -93,5 +93,5 @@ func (*ScrapePerfReplicationGroupMembers) Scrape(ctx context.Context, db *sql.DB
 var scrapePerfReplicationGroupMembers Scraper = &ScrapePerfReplicationGroupMembers{}
 
 func init() {
-	mustRegisterWithDefaults(scrapePerfReplicationGroupMembers, false)
+	mustRegisterScraperWithDefaults(scrapePerfReplicationGroupMembers, false)
 }
