+++ collector/perf_schema_table_io_waits.go
@@ -125,5 +125,5 @@ func (*ScrapePerfTableIOWaits) Scrape(ctx context.Context, db *sql.DB, ch chan<-
 var scrapePerfTableIOWaits Scraper = &ScrapePerfTableIOWaits{}
 
 func init() {
-	mustRegisterWithDefaults(scrapePerfTableIOWaits, false)
+	mustRegisterScraperWithDefaults(scrapePerfTableIOWaits, false)
 }
