+++ collector/slave_status.go
@@ -140,5 +140,5 @@ func (*ScrapeSlaveStatus) Scrape(ctx context.Context, db *sql.DB, ch chan<- prom
 var scrapeSlaveStatus Scraper = &ScrapeSlaveStatus{}
 
 func init() {
-	mustRegisterWithDefaults(scrapeSlaveStatus, true)
+	mustRegisterScraperWithDefaults(scrapeSlaveStatus, true)
 }
