+++ collector/sys_user_summary.go
@@ -156,5 +156,5 @@ func (*ScrapeSysUserSummary) Scrape(ctx context.Context, db *sql.DB, ch chan<- p
 var scrapeSysUserSummary Scraper = &ScrapeSysUserSummary{}
 
 func init() {
-	mustRegisterWithDefaults(scrapeSysUserSummary, false)
+	mustRegisterScraperWithDefaults(scrapeSysUserSummary, false)
 }
